Custom Notification setup for shortlist
