+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.shaan.androiduicomponents"
     xmlns:tools="http://schemas.android.com/tools">
 
+    <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.POST_NOTIFICATIONS" />
 
     <application
@@ -70,6 +72,9 @@
             android:name=".ShortlistActivity"
             android:exported="false"/>
 
+        <activity
+            android:name=".NotificationListActivity"
+            android:exported="false"/>
     </application>
 
 </manifest>
