+++ app/src/main/java/com/shaan/androiduicomponents/models/Deadline.kt
@@ -1,0 +1,6 @@
+package com.shaan.androiduicomponents.models
+data class Deadline(
+    val universityName: String,
+    val type: String,
+    val date: String,
+)
