--- app/src/main/java/com/souravpalitrana/androiduicomponents/Deadline.kt
@@ -1,6 +1,0 @@
-package com.souravpalitrana.androiduicomponents
-data class Deadline(
-    val universityName: String,
-    val type: String,
-    val date: String,
-)
