--- app/src/main/java/com/souravpalitrana/androiduicomponents/MainActivity.kt
@@ -1,27 +1,0 @@
-package com.souravpalitrana.androiduicomponents
-
-import android.os.Bundle
-import android.widget.Toast
-import androidx.appcompat.app.AppCompatActivity
-import net.objecthunter.exp4j.ExpressionBuilder
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_home)
-
-//        setContent {
-//            AndroidUIComponentsTheme {
-//                // A surface container using the 'background' color from the theme
-//                Surface(
-//                    modifier = Modifier.fillMaxSize(),
-//                    color = MaterialTheme.colorScheme.background
-//                ) {
-//                    Greeting("Android")
-//                }
-//            }
-//        }
-    }
-
-
-}
